requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
ftfy==6.3.1
huggingface-hub==0.34.4
idna==3.10
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
open-clip-torch==3.1.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyyaml==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sympy==1.13.1
timm==1.0.19
torch==2.5.1+cu121 # --extra-index-url https://download.pytorch.org/whl/cu121
torchaudio==2.5.1+cu121 # --extra-index-url https://download.pytorch.org/whl/cu121
torchvision==0.20.1+cu121 # --extra-index-url https://download.pytorch.org/whl/cu121
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
wcwidth==0.2.13
